feat(libgraphics): added indirect draw
