+++ libgraphics/include/pivot/graphics/types/RenderObject.hxx
@@ -6,5 +6,5 @@
 struct RenderObject {
     std::string meshID;
     std::string objectName = meshID;
-    gpuObject::UniformBufferObject ubo;
+    ObjectInformation objectInformation;
 };
